Check Maria

+++ sw/airborne/modules/trajectory/trajectory.h
@@ -40,7 +40,7 @@ extern float TRAJECTORY_Y;
 extern float TRAJECTORY_SWITCHING_TIME;
 
 extern float *GLOBAL_OF_VECTOR;
-extern float *GLOBAL_OBJECTS_VECTOR;
+extern int *GLOBAL_OBJECTS_VECTOR;
 
 // functions
 extern void trajectory_init();
